+++ Descriptors/ProtobufDescriptor.h
@@ -26,15 +26,16 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
 
-#ifndef PBF_DESCRIPTOR_H
-#define PBF_DESCRIPTOR_H
+#ifndef PROTOBUF_DESCRIPTOR_H
+#define PROTOBUF_DESCRIPTOR_H
 
-#include "Descriptor.h"
+#include "BaseDescriptor.h"
 #include "response.pb.h"
 
-template <class DataFacadeT> class PBFDescriptor : public Descriptor<DataFacadeT>
+template <class DataFacadeT> class PBFDescriptor : public BaseDescriptor<DataFacadeT>
 {
   private:
+    typedef BaseDescriptor<DataFacadeT> super;
 
     inline void AddInstructionToRoute(protobufResponse::Route &route,
                                       const std::string &id,
@@ -61,17 +62,17 @@ template <class DataFacadeT> class PBFDescriptor : public Descriptor<DataFacadeT
 
     inline void BuildTextualDescription(const int route_length,
                                         DescriptionFactory &description_factory,
-                                        std::vector<typename Descriptor<DataFacadeT>::Segment> &route_segments_list,
+                                        std::vector<typename super::Segment> &route_segments_list,
                                         protobufResponse::Route &route
                                         )
     {
-        std::vector<typename Descriptor<DataFacadeT>::Instruction> instructions;
-        Descriptor<DataFacadeT>::BuildTextualDescription(description_factory,
+        std::vector<typename super::Instruction> instructions;
+        super::BuildTextualDescription(description_factory,
                                                          instructions,
                                                          route_length,
                                                          route_segments_list);
 
-        for (const typename Descriptor<DataFacadeT>::Instruction &i : instructions)
+        for (const typename super::Instruction &i : instructions)
         {
             AddInstructionToRoute(route, i.instructionId, i.streetName, i.length, 
                                   i.position, i.time, i.lengthStr, i.earthDirection,
@@ -80,7 +81,7 @@ template <class DataFacadeT> class PBFDescriptor : public Descriptor<DataFacadeT
     }
 
   public:
-    PBFDescriptor(DataFacadeT *facade) : Descriptor<DataFacadeT>(facade)
+    PBFDescriptor(DataFacadeT *facade) : super(facade)
     {}
 
     void Run(const RawRouteData &raw_route, http::Reply &reply)
@@ -98,13 +99,13 @@ template <class DataFacadeT> class PBFDescriptor : public Descriptor<DataFacadeT
             return;
         }
 
-        std::string road_name = Descriptor<DataFacadeT>::facade->GetEscapedNameForNameID(
+        std::string road_name = super::facade->GetEscapedNameForNameID(
             raw_route.segment_end_coordinates.front().source_phantom.name_id);
 
         BOOST_ASSERT(raw_route.unpacked_path_segments.size() ==
                      raw_route.segment_end_coordinates.size());
 
-        Descriptor<DataFacadeT>::description_factory.SetStartSegment(
+        super::description_factory.SetStartSegment(
             raw_route.segment_end_coordinates.front().source_phantom,
             raw_route.source_traversed_in_reverse.front());
         response.set_status(0);
@@ -116,7 +117,7 @@ template <class DataFacadeT> class PBFDescriptor : public Descriptor<DataFacadeT
 #ifndef NDEBUG
             const int added_segments =
 #endif
-            Descriptor<DataFacadeT>::DescribeLeg(raw_route.unpacked_path_segments[i],
+            super::DescribeLeg(raw_route.unpacked_path_segments[i],
                         raw_route.segment_end_coordinates[i],
                         raw_route.target_traversed_in_reverse[i],
                         raw_route.is_via_leg(i));
@@ -124,30 +125,30 @@ template <class DataFacadeT> class PBFDescriptor : public Descriptor<DataFacadeT
         }
 
         protobufResponse::Route mainRoute;
-        Descriptor<DataFacadeT>::description_factory.Run(Descriptor<DataFacadeT>::facade, Descriptor<DataFacadeT>::config.zoom_level);
-        if (Descriptor<DataFacadeT>::config.geometry)
+        super::description_factory.Run(super::facade, super::config.zoom_level);
+        if (super::config.geometry)
         {
             std::string route_geometry;
-            Descriptor<DataFacadeT>::description_factory.AppendEncodedPolylineStringEncoded(route_geometry);
+            route_geometry = super::description_factory.AppendEncodedPolylineStringEncoded();
             mainRoute.set_route_geometry(route_geometry);
         }
-        if (Descriptor<DataFacadeT>::config.instructions)
+        if (super::config.instructions)
         {
             BuildTextualDescription(raw_route.shortest_path_length,
-                                    Descriptor<DataFacadeT>::description_factory,
-                                    Descriptor<DataFacadeT>::shortest_path_segments,
+                                    super::description_factory,
+                                    super::shortest_path_segments,
                                     mainRoute);
         }
 
-        Descriptor<DataFacadeT>::description_factory.BuildRouteSummary(Descriptor<DataFacadeT>::description_factory.entireLength,
+        super::description_factory.BuildRouteSummary(super::description_factory.entireLength,
                                               raw_route.shortest_path_length);
 
         protobufResponse::RouteSummary routeSummary;
 
-        routeSummary.set_total_distance(Descriptor<DataFacadeT>::description_factory.summary.distance);
-        routeSummary.set_total_time(Descriptor<DataFacadeT>::description_factory.summary.duration);
-        routeSummary.set_start_point(Descriptor<DataFacadeT>::facade->GetEscapedNameForNameID(Descriptor<DataFacadeT>::description_factory.summary.source_name_id));
-        routeSummary.set_end_point(Descriptor<DataFacadeT>::facade->GetEscapedNameForNameID(Descriptor<DataFacadeT>::description_factory.summary.target_name_id));
+        routeSummary.set_total_distance(super::description_factory.summary.distance);
+        routeSummary.set_total_time(super::description_factory.summary.duration);
+        routeSummary.set_start_point(super::facade->GetEscapedNameForNameID(super::description_factory.summary.source_name_id));
+        routeSummary.set_end_point(super::facade->GetEscapedNameForNameID(super::description_factory.summary.target_name_id));
         mainRoute.mutable_route_summary()->CopyFrom(routeSummary);
 
         BOOST_ASSERT(!raw_route.segment_end_coordinates.empty());
@@ -169,62 +170,62 @@ template <class DataFacadeT> class PBFDescriptor : public Descriptor<DataFacadeT
             mainRoute.add_via_points()->CopyFrom(point);
         }
 
-        std::vector<unsigned> const &shortest_leg_end_indices = Descriptor<DataFacadeT>::description_factory.GetViaIndices();
+        std::vector<unsigned> const &shortest_leg_end_indices = super::description_factory.GetViaIndices();
         for (unsigned v : shortest_leg_end_indices)
         {
             mainRoute.add_via_indices(v);
         }
 
         RouteNames route_names =
-            Descriptor<DataFacadeT>::GenerateRouteNames(Descriptor<DataFacadeT>::shortest_path_segments, 
-                Descriptor<DataFacadeT>::alternative_path_segments, Descriptor<DataFacadeT>::facade);
+            super::GenerateRouteNames(super::shortest_path_segments,
+                super::alternative_path_segments, super::facade);
 
         if (INVALID_EDGE_WEIGHT != raw_route.alternative_path_length)
         {
             protobufResponse::Route alternativeRoute;
             BOOST_ASSERT(!raw_route.alt_source_traversed_in_reverse.empty());
-            Descriptor<DataFacadeT>::alternate_description_factory.SetStartSegment(
+            super::alternate_description_factory.SetStartSegment(
                 raw_route.segment_end_coordinates.front().source_phantom,
                 raw_route.alt_source_traversed_in_reverse.front());
             // Get all the coordinates for the computed route
             for (const PathData &path_data : raw_route.unpacked_alternative)
             {
-                Descriptor<DataFacadeT>::current = Descriptor<DataFacadeT>::facade->GetCoordinateOfNode(path_data.node);
-                Descriptor<DataFacadeT>::alternate_description_factory.AppendSegment(Descriptor<DataFacadeT>::current, path_data);
+                super::current = super::facade->GetCoordinateOfNode(path_data.node);
+                super::alternate_description_factory.AppendSegment(super::current, path_data);
             }
-            Descriptor<DataFacadeT>::alternate_description_factory.SetEndSegment(raw_route.segment_end_coordinates.back().target_phantom, raw_route.alt_source_traversed_in_reverse.back());
-            Descriptor<DataFacadeT>::alternate_description_factory.Run(Descriptor<DataFacadeT>::facade, Descriptor<DataFacadeT>::config.zoom_level);
+            super::alternate_description_factory.SetEndSegment(raw_route.segment_end_coordinates.back().target_phantom, raw_route.alt_source_traversed_in_reverse.back());
+            super::alternate_description_factory.Run(super::facade, super::config.zoom_level);
 
-            if (Descriptor<DataFacadeT>::config.geometry)
+            if (super::config.geometry)
             {
                 std::string alternateGeometry;
-                Descriptor<DataFacadeT>::alternate_description_factory.AppendEncodedPolylineStringEncoded(alternateGeometry);
+                alternateGeometry = super::alternate_description_factory.AppendEncodedPolylineStringEncoded();
                 alternativeRoute.set_route_geometry(alternateGeometry);
             }
             // Generate instructions for each alternative (simulated here)
-            if (Descriptor<DataFacadeT>::config.instructions)
+            if (super::config.instructions)
             {
                 BuildTextualDescription(raw_route.alternative_path_length,
-                                        Descriptor<DataFacadeT>::alternate_description_factory,
-                                        Descriptor<DataFacadeT>::alternative_path_segments,
+                                        super::alternate_description_factory,
+                                        super::alternative_path_segments,
                                         alternativeRoute);
             }
 
-            Descriptor<DataFacadeT>::alternate_description_factory.BuildRouteSummary(
-                Descriptor<DataFacadeT>::alternate_description_factory.entireLength, raw_route.alternative_path_length);
+            super::alternate_description_factory.BuildRouteSummary(
+                super::alternate_description_factory.entireLength, raw_route.alternative_path_length);
 
             protobufResponse::RouteSummary alternativeRouteSummary;
 
-            alternativeRouteSummary.set_total_distance(Descriptor<DataFacadeT>::alternate_description_factory.summary.distance);
-            alternativeRouteSummary.set_total_time(Descriptor<DataFacadeT>::alternate_description_factory.summary.duration);
-            alternativeRouteSummary.set_start_point(Descriptor<DataFacadeT>::facade->GetEscapedNameForNameID(
-                Descriptor<DataFacadeT>::alternate_description_factory.summary.source_name_id));
-            alternativeRouteSummary.set_end_point(Descriptor<DataFacadeT>::facade->GetEscapedNameForNameID(
-                Descriptor<DataFacadeT>::alternate_description_factory.summary.target_name_id));
+            alternativeRouteSummary.set_total_distance(super::alternate_description_factory.summary.distance);
+            alternativeRouteSummary.set_total_time(super::alternate_description_factory.summary.duration);
+            alternativeRouteSummary.set_start_point(super::facade->GetEscapedNameForNameID(
+                super::alternate_description_factory.summary.source_name_id));
+            alternativeRouteSummary.set_end_point(super::facade->GetEscapedNameForNameID(
+                super::alternate_description_factory.summary.target_name_id));
             alternativeRoute.mutable_route_summary()->CopyFrom(routeSummary);
 
             std::vector<unsigned> const &alternate_leg_end_indices =
-                Descriptor<DataFacadeT>::alternate_description_factory.GetViaIndices();
+                super::alternate_description_factory.GetViaIndices();
             for (unsigned v : alternate_leg_end_indices)
             {
                 alternativeRoute.add_via_indices(v);
@@ -259,5 +260,5 @@ template <class DataFacadeT> class PBFDescriptor : public Descriptor<DataFacadeT
         reply.content.insert(reply.content.end(), output.begin(), output.end());
     }
 };
-#endif // PBF_DESCRIPTOR_H
+#endif // PROTOBUF_DESCRIPTOR_H
